:gear: Implémentation plus simple des dépendants implicites aux parents
Plus simple que be7b2b8ac4e747b6a7cb4d56e2edc8544875f4c4

This feature was at first designed to work not only with boolean questions, but also applicability conditions and formulas. But lots of numeric formulas are also namespaces, with children in the computations themselves. Which, in the previous implementations, lead to lots of irrelevant checks. To be reintroduced better if needed